Move RemoteNode to its own module; move try_query_certificates_from there.
